Keep the same signature when resource exists.
The resource being created is returned, even if it already exists. This
triggers a second HTTP call because there is currently no other way to do it
with the Kinto server: an error is returned with the record that is
conflicting, but no permissions are attached to it, and the create_*resource*
usually returns the permissions *and* the data.